% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayvec.R
\name{print.arrayvec}
\alias{print.arrayvec}
\alias{arrayvec}
\alias{dim.arrayvec}
\alias{as.matrix.arrayvec}
\alias{as.data.frame.arrayvec}
\alias{[.arrayvec}
\alias{[[.arrayvec}
\title{Support for C++ vector of arrays}
\usage{
\method{print}{arrayvec}(x, ...)

\method{dim}{arrayvec}(x)

\method{as.matrix}{arrayvec}(x, ...)

\method{as.data.frame}{arrayvec}(x, ...)

\method{[}{arrayvec}(x, i, j, drop = TRUE)

\method{[[}{arrayvec}(x, ...)
}
\arguments{
\item{x}{an arrayvec object}

\item{...}{other parameters}

\item{i}{row}

\item{j}{column}

\item{drop}{drop singleton dimensions if true}
}
\value{
\tabular{ll}{
\code{print.arrayvec} \tab the object invisibly \cr
\code{dim.arrayvec} \tab the rows and columns \cr
\code{as.matrix.arrayvec} \tab a matrix \cr
\code{as.data.frame.arrayvec} \tab a data frame \cr
\code{`[.arrayvec`} \tab a matrix or vector \cr
\code{`[[.arrayvec`} \tab a column vector \cr
}
}
\description{
Support for C++ vector of arrays
}
\details{
Because \code{kdtools} is implemented in C++, it operates natively
  on a vector of arrays. An \code{arrayvec} object is a wrapper around a
  pointer to a vector of arrays. These functions provide some ability to
  manipulate the data as if it were a matrix.
}
