% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools.R
\name{colspec}
\alias{colspec}
\title{Generate column indices}
\usage{
colspec(x, cols = NULL)
}
\arguments{
\item{x}{a data frame}

\item{cols}{a column specification}
}
\value{
A vector of integers
}
\description{
Generate column indices
}
\details{
\code{cols} can be a logical, numeric, or character vector
indicating which columns to extract. Or \code{cols} can be a formula
with the right-hand-side variable specifying the columns. This function
is used where ever a \code{cols} option is present throughout \code{kdtools}.
}
\examples{
colspec(mtcars, c(1, 3, 6))
colspec(mtcars, c("mpg", "disp", "wt"))
colspec(mtcars, ~mpg + disp + wt)
colspec(mtcars, c(TRUE, FALSE, TRUE, FALSE, FALSE, TRUE,
                  FALSE, FALSE, FALSE, FALSE, FALSE))

}
