% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_to_tuples}
\alias{matrix_to_tuples}
\alias{tuples_to_matrix}
\title{Convert a matrix to a vector of arrays}
\usage{
matrix_to_tuples(x)

tuples_to_matrix(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a matrix to a vector of arrays
}
\details{
The algorithms in kdtools can accept either matrices or an
\link{arrayvec} object. When a matrix is passed, it is converted to
an arrayvec object internally and the results are converted back to
a matrix. For optimal performance, pre-convert matrices.
}
\examples{
if (has_cxx17()) {
x = matrix(1:10, 5)
y = matrix_to_tuples(x)
str(x)
str(y)
y[1:2, ]
}

}
