% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexvec.R
\name{print.indexvec}
\alias{print.indexvec}
\alias{indexvec}
\alias{dim.indexvec}
\alias{as.matrix.indexvec}
\alias{as.data.frame.indexvec}
\alias{[.indexvec}
\alias{[[.indexvec}
\title{Support for C++ vector of arrays}
\usage{
\method{print}{indexvec}(x, ...)

\method{dim}{indexvec}(x)

\method{as.matrix}{indexvec}(x, ...)

\method{as.data.frame}{indexvec}(x, ...)

\method{[}{indexvec}(x, i, j, drop = TRUE)

\method{[[}{indexvec}(x, ...)
}
\arguments{
\item{x}{an indexvec object}

\item{...}{other parameters}

\item{i}{row}

\item{j}{column}

\item{drop}{drop singleton dimensions if true}
}
\description{
Support for C++ vector of arrays
}
\details{
Because \code{kdtools} is implemented in C++, it operates natively
  on a vector of arrays. An \code{indexvec} object is a wrapper around a
  pointer to a vector of arrays. These functions provide some ability to
  manipulate the data as if it were a matrix.
}
