% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{levenshtein}
\alias{levenshtein}
\title{Compute edit distance}
\usage{
levenshtein(s1, s2)
}
\arguments{
\item{s1}{first character vector}

\item{s2}{second character vector}
}
\value{
the integer number of edits between strings
}
\description{
Compute edit distance
}
\details{
If neither string is greater than ten characters, there is
no allocation. Otherwise, allocation is amortized constant.
}
