% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools.R
\name{kd_lower_bound}
\alias{kd_lower_bound}
\alias{kd_upper_bound}
\alias{kd_range_query}
\alias{kd_binary_search}
\title{Search sorted data}
\usage{
kd_lower_bound(x, v)

kd_upper_bound(x, v)

kd_range_query(x, l, u)

kd_binary_search(x, v)
}
\arguments{
\item{x}{an object sorted by \code{\link{kd_sort}}}

\item{v}{a vector specifying where to look}

\item{l}{lower left corner of search region}

\item{u}{upper right corner of search region}
}
\description{
Search sorted data
}
