% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{using_circular_lexicographical_compare}
\alias{using_circular_lexicographical_compare}
\alias{has_cxx17}
\title{Check if package was compiled with circular comparisons}
\usage{
using_circular_lexicographical_compare()

has_cxx17()
}
\description{
Check if package was compiled with circular comparisons

Check if C++ 17 was available when building package
}
