% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_to_tuples}
\alias{matrix_to_tuples}
\alias{tuples_to_matrix}
\title{Convert a matrix to a vector of arrays}
\usage{
matrix_to_tuples(x)

tuples_to_matrix(x)
}
\arguments{
\item{x}{object to be converted}
}
\description{
Convert a matrix to a vector of arrays
}
\details{
Most of the R functions exposed by \code{kdtools} can operate
natively on a data.frame, matrix or an \code{\link{arrayvec}} object.
An \code{arrayvec} object is a pointer to a native C++ vector of arrays
and will be the fastest option. Only doubles are supported by \code{arrayvec}.
Use this function to convert your data before processing for optimal speed.
}
\examples{
if (has_cxx17()) {
x = matrix(1:10, 5)
y = matrix_to_tuples(x)
str(x)
str(y)
y[1:2, ]
}

}
