% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools-package.R
\docType{package}
\name{kdtools-package}
\alias{kdtools}
\alias{kdtools-package}
\title{kdtools: Fast Sorting and Searching of Multidimensional Data}
\description{
Sorted data can be searched in logarithmic time using divide-and-conquer schemes. Here, we extend this idea to k-dimensional data to allow rapid range-queries and exact and appriximate nearest-neighbor queries on data stored in matrices and data frames. A variant of the quick-sort algorithm rotates among dimensions to produce an implicit k-d tree ordering of the data. The package includes 'C++' code that can be used independently of R. Representation of a binary tree in a sorted list is a common technique in computer science \url{https://en.wikipedia.org/wiki/Implicit_data_structure}. The original presentation of k-d trees can be found in Bentley (1975) \doi{10.1145/361002.361007}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thk686/kdtools}
  \item Report bugs at \url{https://github.com/thk686/kdtools/issues}
}

}
\author{
\strong{Maintainer}: Timothy Keitt \email{tkeitt@gmail.com}

}
\keyword{internal}
