% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools.R
\name{lex_sort}
\alias{lex_sort}
\title{Sort a matrix into lexicographical order}
\usage{
lex_sort(x, ...)
}
\arguments{
\item{x}{a matrix or arrayvec object}

\item{...}{other parameters}
}
\value{
the input type sorted
}
\description{
Sort a matrix into lexicographical order
}
\details{
Sorts a range of tuples into lexicographical order. This function
  only exists for demonstration purposes.
}
\examples{
if (has_cxx17()) {
x = lex_sort(matrix(runif(200), 100))
plot(x, type = "o", pch = 19, col = "steelblue", asp = 1)
}
}
