% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools.R
\name{kd_nearest_neighbors}
\alias{kd_nearest_neighbors}
\alias{kd_nearest_neighbor}
\alias{kd_approx_nn}
\title{Find nearest neighbors}
\usage{
kd_nearest_neighbors(x, v, n)

kd_nearest_neighbor(x, v)

kd_approx_nn(x, v, eps)
}
\arguments{
\item{x}{an object sorted by \code{\link{kd_sort}}}

\item{v}{a vector specifying where to look}

\item{n}{the number of neighbors to return}

\item{eps}{maximum allowable distance error}
}
\description{
Find nearest neighbors
}
