% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtools.R
\name{kd_nearest_neighbors}
\alias{kd_nearest_neighbors}
\alias{kd_nearest_neighbor}
\title{Find nearest neighbors}
\usage{
kd_nearest_neighbors(x, v, n)

kd_nearest_neighbor(x, v)
}
\arguments{
\item{x}{an object sorted by \code{\link{kd_sort}}}

\item{v}{a vector specifying where to look}

\item{n}{the number of neighbors to return}
}
\description{
Find nearest neighbors
}
\examples{
x = matrix(runif(200), 100)
y = matrix_to_tuples(x)
kd_sort(y, inplace = TRUE)
y[kd_nearest_neighbor(y, c(1/2, 1/2)),]
kd_nearest_neighbors(y, c(1/2, 1/2), 3)

}
